Struct vncserver::_rfbClientRec

#[repr(C)]pub struct _rfbClientRec {
    pub screen: rfbScreenInfoPtr,
    pub scaledScreen: rfbScreenInfoPtr,
    pub PalmVNC: rfbBool,
    pub clientData: *mut c_void,
    pub clientGoneHook: ClientGoneHookPtr,
    pub sock: c_int,
    pub host: *mut c_char,
    pub protocolMajorVersion: c_int,
    pub protocolMinorVersion: c_int,
    pub client_thread: pthread_t,
    pub state: _rfbClientRec__bindgen_ty_1,
    pub reverseConnection: rfbBool,
    pub onHold: rfbBool,
    pub readyForSetColourMapEntries: rfbBool,
    pub useCopyRect: rfbBool,
    pub preferredEncoding: c_int,
    pub correMaxWidth: c_int,
    pub correMaxHeight: c_int,
    pub viewOnly: rfbBool,
    pub authChallenge: [u8; 16],
    pub copyRegion: sraRegionPtr,
    pub copyDX: c_int,
    pub copyDY: c_int,
    pub modifiedRegion: sraRegionPtr,
    pub requestedRegion: sraRegionPtr,
    pub startDeferring: timeval,
    pub startPtrDeferring: timeval,
    pub lastPtrX: c_int,
    pub lastPtrY: c_int,
    pub lastPtrButtons: c_int,
    pub translateFn: rfbTranslateFnType,
    pub translateLookupTable: *mut c_char,
    pub format: rfbPixelFormat,
    pub updateBuf: [c_char; 30000],
    pub ublen: c_int,
    pub statEncList: *mut _rfbStatList,
    pub statMsgList: *mut _rfbStatList,
    pub rawBytesEquivalent: c_int,
    pub bytesSent: c_int,
    pub compStream: z_stream_s,
    pub compStreamInited: rfbBool,
    pub zlibCompressLevel: u32,
    pub tightQualityLevel: c_int,
    pub zsStruct: [z_stream; 4],
    pub zsActive: [rfbBool; 4],
    pub zsLevel: [c_int; 4],
    pub tightCompressLevel: c_int,
    pub compStreamInitedLZO: rfbBool,
    pub lzoWrkMem: *mut c_char,
    pub fileTransfer: rfbFileTransferData,
    pub lastKeyboardLedState: c_int,
    pub enableSupportedMessages: rfbBool,
    pub enableSupportedEncodings: rfbBool,
    pub enableServerIdentity: rfbBool,
    pub enableKeyboardLedState: rfbBool,
    pub enableLastRectEncoding: rfbBool,
    pub enableCursorShapeUpdates: rfbBool,
    pub enableCursorPosUpdates: rfbBool,
    pub useRichCursorEncoding: rfbBool,
    pub cursorWasChanged: rfbBool,
    pub cursorWasMoved: rfbBool,
    pub cursorX: c_int,
    pub cursorY: c_int,
    pub useNewFBSize: rfbBool,
    pub newFBSizePending: rfbBool,
    pub prev: *mut _rfbClientRec,
    pub next: *mut _rfbClientRec,
    pub refCount: c_int,
    pub refCountMutex: pthread_mutex_t,
    pub deleteCond: pthread_cond_t,
    pub outputMutex: pthread_mutex_t,
    pub updateMutex: pthread_mutex_t,
    pub updateCond: pthread_cond_t,
    pub zrleData: *mut c_void,
    pub zywrleLevel: c_int,
    pub zywrleBuf: [c_int; 4096],
    pub progressiveSliceY: c_int,
    pub extensions: *mut rfbExtensionData,
    pub zrleBeforeBuf: *mut c_char,
    pub paletteHelper: *mut c_void,
    pub sendMutex: pthread_mutex_t,
    pub beforeEncBuf: *mut c_char,
    pub beforeEncBufSize: c_int,
    pub afterEncBuf: *mut c_char,
    pub afterEncBufSize: c_int,
    pub afterEncBufLen: c_int,
    pub tightEncoding: u32,
    pub turboSubsampLevel: c_int,
    pub turboQualityLevel: c_int,
    pub sslctx: *mut rfbSslCtx,
    pub wsctx: *mut wsCtx,
    pub wspath: *mut c_char,
    pub readFromSocket: ClientReadFromSocket,
    pub peekAtSocket: ClientPeekAtSocket,
    pub hasPendingOnSocket: ClientHasPendingOnSocket,
    pub writeToSocket: ClientWriteToSocket,
}

Fields
screen: rfbScreenInfoPtr

back pointer to the screen

scaledScreen: rfbScreenInfoPtr

points to a scaled version of the screen buffer in cl->scaledScreenList

PalmVNC: rfbBool

how did the client tell us it wanted the screen changed? Ultra style or palm style?

clientData: *mut c_void

private data. You should put any application client specific data into a struct and let clientData point to it. Don’t forget to free the struct via clientGoneHook!

This is useful if the IO functions have to behave client specific.

clientGoneHook: ClientGoneHookPtrsock: c_inthost: *mut c_charprotocolMajorVersion: c_intprotocolMinorVersion: c_intclient_thread: pthread_tstate: _rfbClientRec__bindgen_ty_1reverseConnection: rfbBoolonHold: rfbBoolreadyForSetColourMapEntries: rfbBooluseCopyRect: rfbBoolpreferredEncoding: c_intcorreMaxWidth: c_intcorreMaxHeight: c_intviewOnly: rfbBoolauthChallenge: [u8; 16]copyRegion: sraRegionPtr

< the destination region of the copy

copyDX: c_int

< the translation by which the copy happens

copyDY: c_int

< the translation by which the copy happens

modifiedRegion: sraRegionPtrrequestedRegion: sraRegionPtr

As part of the FramebufferUpdateRequest, a client can express interest in a subrectangle of the whole framebuffer. This is stored in the requestedRegion member. In the normal case this is the whole framebuffer if the client is ready, empty if it’s not.

startDeferring: timeval

The following member represents the state of the “deferred update” timer

when the framebuffer is modified and the client is ready, in most cases it is more efficient to defer sending the update by a few milliseconds so that several changes to the framebuffer can be combined into a single update.

startPtrDeferring: timevallastPtrX: c_intlastPtrY: c_intlastPtrButtons: c_inttranslateFn: rfbTranslateFnType

translateFn points to the translation function which is used to copy and translate a rectangle from the framebuffer to an output buffer.

translateLookupTable: *mut c_charformat: rfbPixelFormatupdateBuf: [c_char; 30000]ublen: c_intstatEncList: *mut _rfbStatListstatMsgList: *mut _rfbStatListrawBytesEquivalent: c_intbytesSent: c_intcompStream: z_stream_scompStreamInited: rfbBoolzlibCompressLevel: u32tightQualityLevel: c_int

the quality level is also used by ZYWRLE and TightPng

zsStruct: [z_stream; 4]zsActive: [rfbBool; 4]zsLevel: [c_int; 4]tightCompressLevel: c_intcompStreamInitedLZO: rfbBoollzoWrkMem: *mut c_charfileTransfer: rfbFileTransferDatalastKeyboardLedState: c_int

< keep track of last value so we can send change events

enableSupportedMessages: rfbBool

< client supports SupportedMessages encoding

enableSupportedEncodings: rfbBool

< client supports SupportedEncodings encoding

enableServerIdentity: rfbBool

< client supports ServerIdentity encoding

enableKeyboardLedState: rfbBool

< client supports KeyboardState encoding

enableLastRectEncoding: rfbBool

< client supports LastRect encoding

enableCursorShapeUpdates: rfbBool

< client supports cursor shape updates

enableCursorPosUpdates: rfbBool

< client supports cursor position updates

useRichCursorEncoding: rfbBool

< rfbEncodingRichCursor is preferred

cursorWasChanged: rfbBool

< cursor shape update should be sent

cursorWasMoved: rfbBool

< cursor position update should be sent

cursorX: c_int

< the coordinates of the cursor, if enableCursorShapeUpdates = FALSE

cursorY: c_int

< the coordinates of the cursor, if enableCursorShapeUpdates = FALSE

useNewFBSize: rfbBool

< client supports NewFBSize encoding

newFBSizePending: rfbBool

< framebuffer size was changed

prev: *mut _rfbClientRecnext: *mut _rfbClientRecrefCount: c_int

whenever a client is referenced, the refCount has to be incremented and afterwards decremented, so that the client is not cleaned up while being referenced. Use the functions rfbIncrClientRef(cl) and rfbDecrClientRef(cl);

refCountMutex: pthread_mutex_tdeleteCond: pthread_cond_toutputMutex: pthread_mutex_tupdateMutex: pthread_mutex_tupdateCond: pthread_cond_tzrleData: *mut c_voidzywrleLevel: c_intzywrleBuf: [c_int; 4096]progressiveSliceY: c_int

if progressive updating is on, this variable holds the current y coordinate of the progressive slice.

extensions: *mut rfbExtensionDatazrleBeforeBuf: *mut c_char

for threaded zrle

paletteHelper: *mut c_voidsendMutex: pthread_mutex_tbeforeEncBuf: *mut c_charbeforeEncBufSize: c_intafterEncBuf: *mut c_charafterEncBufSize: c_intafterEncBufLen: c_inttightEncoding: u32turboSubsampLevel: c_intturboQualityLevel: c_intsslctx: *mut rfbSslCtxwsctx: *mut wsCtxwspath: *mut c_charreadFromSocket: ClientReadFromSocketpeekAtSocket: ClientPeekAtSockethasPendingOnSocket: ClientHasPendingOnSocketwriteToSocket: ClientWriteToSocket