Struct vncserver::_rfbScreenInfo

#[repr(C)]pub struct _rfbScreenInfo {
    pub scaledScreenNext: *mut _rfbScreenInfo,
    pub scaledScreenRefCount: c_int,
    pub width: c_int,
    pub paddedWidthInBytes: c_int,
    pub height: c_int,
    pub depth: c_int,
    pub bitsPerPixel: c_int,
    pub sizeInBytes: c_int,
    pub blackPixel: rfbPixel,
    pub whitePixel: rfbPixel,
    pub screenData: *mut c_void,
    pub serverFormat: rfbPixelFormat,
    pub colourMap: rfbColourMap,
    pub desktopName: *const c_char,
    pub thisHost: [c_char; 255],
    pub autoPort: rfbBool,
    pub port: c_int,
    pub listenSock: c_int,
    pub maxSock: c_int,
    pub maxFd: c_int,
    pub allFds: fd_set,
    pub socketState: rfbSocketState,
    pub inetdSock: c_int,
    pub inetdInitDone: rfbBool,
    pub udpPort: c_int,
    pub udpSock: c_int,
    pub udpClient: *mut _rfbClientRec,
    pub udpSockConnected: rfbBool,
    pub udpRemoteAddr: sockaddr_in,
    pub maxClientWait: c_int,
    pub httpInitDone: rfbBool,
    pub httpEnableProxyConnect: rfbBool,
    pub httpPort: c_int,
    pub httpDir: *mut c_char,
    pub httpListenSock: c_int,
    pub httpSock: c_int,
    pub passwordCheck: rfbPasswordCheckProcPtr,
    pub authPasswdData: *mut c_void,
    pub authPasswdFirstViewOnly: c_int,
    pub maxRectsPerUpdate: c_int,
    pub deferUpdateTime: c_int,
    pub alwaysShared: rfbBool,
    pub neverShared: rfbBool,
    pub dontDisconnect: rfbBool,
    pub clientHead: *mut _rfbClientRec,
    pub pointerClient: *mut _rfbClientRec,
    pub cursorX: c_int,
    pub cursorY: c_int,
    pub underCursorBufferLen: c_int,
    pub underCursorBuffer: *mut c_char,
    pub dontConvertRichCursorToXCursor: rfbBool,
    pub cursor: *mut rfbCursor,
    pub frameBuffer: *mut c_char,
    pub kbdAddEvent: rfbKbdAddEventProcPtr,
    pub kbdReleaseAllKeys: rfbKbdReleaseAllKeysProcPtr,
    pub ptrAddEvent: rfbPtrAddEventProcPtr,
    pub setXCutText: rfbSetXCutTextProcPtr,
    pub getCursorPtr: rfbGetCursorProcPtr,
    pub setTranslateFunction: rfbSetTranslateFunctionProcPtr,
    pub setSingleWindow: rfbSetSingleWindowProcPtr,
    pub setServerInput: rfbSetServerInputProcPtr,
    pub getFileTransferPermission: rfbFileTransferPermitted,
    pub setTextChat: rfbSetTextChat,
    pub newClientHook: rfbNewClientHookPtr,
    pub displayHook: rfbDisplayHookPtr,
    pub getKeyboardLedStateHook: rfbGetKeyboardLedStateHookPtr,
    pub cursorMutex: pthread_mutex_t,
    pub backgroundLoop: rfbBool,
    pub ignoreSIGPIPE: rfbBool,
    pub progressiveSliceHeight: c_int,
    pub listenInterface: in_addr_t,
    pub deferPtrUpdateTime: c_int,
    pub handleEventsEagerly: rfbBool,
    pub versionString: *mut c_char,
    pub protocolMajorVersion: c_int,
    pub protocolMinorVersion: c_int,
    pub permitFileTransfer: rfbBool,
    pub displayFinishedHook: rfbDisplayFinishedHookPtr,
    pub xvpHook: rfbXvpHookPtr,
    pub sslkeyfile: *mut c_char,
    pub sslcertfile: *mut c_char,
    pub ipv6port: c_int,
    pub listen6Interface: *mut c_char,
    pub listen6Sock: c_int,
    pub http6Port: c_int,
    pub httpListen6Sock: c_int,
}

Per-screen (framebuffer) structure. There can be as many as you wish, each serving different clients. However, you have to call rfbProcessEvents for each of these.

Fields
scaledScreenNext: *mut _rfbScreenInfo

this structure has children that are scaled versions of this screen

scaledScreenRefCount: c_intwidth: c_intpaddedWidthInBytes: c_intheight: c_intdepth: c_intbitsPerPixel: c_intsizeInBytes: c_intblackPixel: rfbPixelwhitePixel: rfbPixelscreenData: *mut c_void

some screen specific data can be put into a struct where screenData points to. You need this if you have more than one screen at the same time while using the same functions.

serverFormat: rfbPixelFormatcolourMap: rfbColourMap

< set this if rfbServerFormat.trueColour==FALSE

desktopName: *const c_charthisHost: [c_char; 255]autoPort: rfbBoolport: c_intlistenSock: c_intmaxSock: c_intmaxFd: c_intallFds: fd_setsocketState: rfbSocketStateinetdSock: c_intinetdInitDone: rfbBooludpPort: c_intudpSock: c_intudpClient: *mut _rfbClientRecudpSockConnected: rfbBooludpRemoteAddr: sockaddr_inmaxClientWait: c_inthttpInitDone: rfbBoolhttpEnableProxyConnect: rfbBoolhttpPort: c_inthttpDir: *mut c_charhttpListenSock: c_inthttpSock: c_intpasswordCheck: rfbPasswordCheckProcPtrauthPasswdData: *mut c_voidauthPasswdFirstViewOnly: c_int

If rfbAuthPasswdData is given a list, this is the first view only password.

maxRectsPerUpdate: c_int

send only this many rectangles in one update

deferUpdateTime: c_int

this is the amount of milliseconds to wait at least before sending an update.

alwaysShared: rfbBoolneverShared: rfbBooldontDisconnect: rfbBoolclientHead: *mut _rfbClientRecpointerClient: *mut _rfbClientRec

< “Mutex” for pointer events

cursorX: c_intcursorY: c_intunderCursorBufferLen: c_intunderCursorBuffer: *mut c_chardontConvertRichCursorToXCursor: rfbBoolcursor: *mut rfbCursorframeBuffer: *mut c_char

the frameBuffer has to be supplied by the serving process. The buffer will not be freed by

kbdAddEvent: rfbKbdAddEventProcPtrkbdReleaseAllKeys: rfbKbdReleaseAllKeysProcPtrptrAddEvent: rfbPtrAddEventProcPtrsetXCutText: rfbSetXCutTextProcPtrgetCursorPtr: rfbGetCursorProcPtrsetTranslateFunction: rfbSetTranslateFunctionProcPtrsetSingleWindow: rfbSetSingleWindowProcPtrsetServerInput: rfbSetServerInputProcPtrgetFileTransferPermission: rfbFileTransferPermittedsetTextChat: rfbSetTextChatnewClientHook: rfbNewClientHookPtr

newClientHook is called just after a new client is created

displayHook: rfbDisplayHookPtr

displayHook is called just before a frame buffer update

getKeyboardLedStateHook: rfbGetKeyboardLedStateHookPtr

These hooks are called to pass keyboard state back to the client

cursorMutex: pthread_mutex_tbackgroundLoop: rfbBoolignoreSIGPIPE: rfbBool

if TRUE, an ignoring signal handler is installed for SIGPIPE

progressiveSliceHeight: c_int

if not zero, only a slice of this height is processed every time an update should be sent. This should make working on a slow link more interactive.

listenInterface: in_addr_tdeferPtrUpdateTime: c_inthandleEventsEagerly: rfbBool

handle as many input events as possible (default off)

versionString: *mut c_char

rfbEncodingServerIdentity

protocolMajorVersion: c_int

What does the server tell the new clients which version it supports

protocolMinorVersion: c_intpermitFileTransfer: rfbBool

command line authorization of file transfers

displayFinishedHook: rfbDisplayFinishedHookPtr

displayFinishedHook is called just after a frame buffer update

xvpHook: rfbXvpHookPtr

xvpHook is called to handle an xvp client message

sslkeyfile: *mut c_charsslcertfile: *mut c_charipv6port: c_int

< The port to listen on when using IPv6.

listen6Interface: *mut c_charlisten6Sock: c_inthttp6Port: c_inthttpListen6Sock: c_int